refactor: implement ConfigProvider and useConfig hook for centralized configuration management
