+++ src/features/config/index.ts
@@ -1,0 +1,2 @@
+export { default as ConfigProvider } from './provider'
+export { default as useConfig } from './hook'
